btrfs-progs: reduce size of btrfs_path::lowest_level

+++ ctree.h
@@ -561,7 +561,7 @@ struct btrfs_path {
 #endif
 	signed char reada;
 	/* keep some upper locks as we walk down */
-	int lowest_level;
+	u8 lowest_level;
 
 	/*
 	 * set by btrfs_split_item, tells search_slot to keep all locks
